Make an items package. Move item_manager to it. Make rocks.py and move all rock related classes there. This will prevent circular imports.
